Moves logic and responsibility to a higher level for the incident-management module
