+++ src/signals/incident-management/constants.js
@@ -1,0 +1,9 @@
+export const REMOVE_FILTER = 'sia/IncidentOverviewPage/REMOVE_FILTER';
+export const REMOVE_FILTER_SUCCESS = 'sia/IncidentOverviewPage/REMOVE_FILTER_SUCCESS';
+export const REMOVE_FILTER_FAILED = 'sia/IncidentOverviewPage/REMOVE_FILTER_FAILED';
+
+export const APPLY_FILTER = 'sia/IncidentOverviewPage/APPLY_FILTER';
+
+export const GET_FILTERS = 'sia/incidentManagement/GET_FILTERS';
+export const GET_FILTERS_SUCCESS = 'sia/incidentManagement/GET_FILTERS_SUCCESS';
+export const GET_FILTERS_FAILED = 'sia/incidentManagement/GET_FILTERS_FAILED';
