+++ src/signals/incident-management/containers/Filter/index.js
@@ -12,9 +12,8 @@ import {
   requestIncidents,
   incidentSelected as onIncidentSelected,
 } from 'signals/incident-management/containers/IncidentOverviewPage/actions';
-import makeSelectOverviewPage, {
-  makeSelectFilter,
-} from 'signals/incident-management/containers/IncidentOverviewPage/selectors';
+import makeSelectOverviewPage from 'signals/incident-management/containers/IncidentOverviewPage/selectors';
+import { makeSelectFilter } from 'signals/incident-management/selectors';
 import FilterForm from 'signals/incident-management/components/FilterForm';
 import { resetSearchQuery } from 'models/search/actions';
