+++ src/signals/incident-management/containers/IncidentOverviewPage/constants.js
@@ -7,14 +7,5 @@ export const FILTER_INCIDENTS_CHANGED = 'sia/IncidentOverviewPage/FILTER_INCIDEN
 export const PAGE_INCIDENTS_CHANGED = 'sia/IncidentOverviewPage/PAGE_INCIDENTS_CHANGED';
 export const SORT_INCIDENTS_CHANGED = 'sia/IncidentOverviewPage/SORT_INCIDENTS_CHANGED';
 
-export const GET_FILTERS = 'sia/IncidentOverviewPage/GET_FILTERS';
-export const GET_FILTERS_SUCCESS = 'sia/IncidentOverviewPage/GET_FILTERS_SUCCESS';
-export const GET_FILTERS_FAILED = 'sia/IncidentOverviewPage/GET_FILTERS_FAILED';
-
-export const REMOVE_FILTER = 'sia/IncidentOverviewPage/REMOVE_FILTER';
-export const REMOVE_FILTER_SUCCESS = 'sia/IncidentOverviewPage/REMOVE_FILTER_SUCCESS';
-export const REMOVE_FILTER_FAILED = 'sia/IncidentOverviewPage/REMOVE_FILTER_FAILED';
-
-export const APPLY_FILTER = 'sia/IncidentOverviewPage/APPLY_FILTER';
 export const APPLY_FILTER_REFRESH = 'sia/IncidentOverviewPage/APPLY_FILTER_REFRESH';
 export const APPLY_FILTER_REFRESH_STOP = 'sia/IncidentOverviewPage/APPLY_FILTER_REFRESH_STOP';
