+++ src/signals/incident-management/containers/IncidentOverviewPage/selectors.js
@@ -32,50 +32,5 @@ const makeSelectIncidentsCount = createSelector(
   },
 );
 
-const makeSelectFilterParams = () =>
-  createSelector(
-    selectOverviewPageDomain,
-    (substate) => {
-      const state = substate.toJS();
-      const filter = state.filter || { options: {} };
-      const { options } = filter;
-      if (options && options.id) {
-        delete options.id;
-      }
-
-      if (filter.searchQuery) {
-        return {
-          id: filter.searchQuery,
-          page: state.page,
-          ordering: state.sort,
-        };
-      }
-
-      return { ...options, page: state.page, ordering: state.sort };
-    },
-  );
-
 export default makeSelectOverviewPage;
-export {
-  selectOverviewPageDomain,
-  makeSelectFilterParams,
-  makeSelectIncidentsCount,
-};
-
-export const makeSelectAllFilters = createSelector(
-  selectOverviewPageDomain,
-  (stateMap) => {
-    const state = stateMap.toJS();
-
-    return state.allFilters;
-  },
-);
-
-export const makeSelectFilter = createSelector(
-  selectOverviewPageDomain,
-  (stateMap) => {
-    const state = stateMap.toJS();
-
-    return state.filter;
-  },
-);
+export { selectOverviewPageDomain, makeSelectIncidentsCount };
